harden for production editors: OOM injection, threads, fuzz, CI
- Point.row/col widen u32 -> usize: a >4 GiB single line must not
  truncate (release) or panic (safe); Point is a transient value type,
  not tree storage, so this costs nothing.
- replace is now atomic under OOM (pre-replace root retained; delete/
  insert restore themselves) and gains a single-leaf in-place fast path.
- insert general path: build the middle tree BEFORE consuming the root.
  Found by allocation-fault injection: the old order leaked the retained
  root (and thus the tree) when the middle build failed.
- tests: checkAllAllocationFailures over a scripted general-path sequence
  (all 76 allocation sites; each induced failure must leave rope AND live
  snapshot byte-identical, leak-free); 6-thread snapshot-verification
  under live edits (smp_allocator); Smith-driven fuzz op stream vs the
  oracle (--fuzz-able); 8 MiB borrowed-backing doc with cross-span edits.
- CI: fmt gate + 3-mode tests + bench-build smoke + nix-build -A default,
  toolchain from the npins pin (snail's preamble). nix-build verified
  locally end to end.

+++ src/stemma/geometry.zig
@@ -24,10 +24,11 @@ pub const Range = struct {
 
 /// A line/column position. `col` is a **byte** offset within the line — not a
 /// grapheme or display column (those are the caller's job). Row and column are
-/// 0-based; the newline terminating a row is not part of the row.
+/// 0-based; the newline terminating a row is not part of the row. Full-width
+/// integers on purpose: a multi-gigabyte single-line file must not truncate.
 pub const Point = struct {
-    row: u32,
-    col: u32,
+    row: usize,
+    col: usize,
 };
 
 /// Tie-breaking side for an `Anchor` sitting exactly at an edit boundary.
